Fix crash in banded_ssw
The loop for banded_sw sometimes cannot achieve the maximum score so it just keeps doubling the bandwidth size indefinitely until the program crashes.  I'm not sure what the best fix is here but I simply bounded the loop as follows:
The original loop in banded_sw looks like
do {
..
} while (LIKELY(max < score);

I changed this to 
do {
} while (LIKELY(max < score) && LIKELY(band_width <= readLen));

As I understand it, band_width > readLen doesn't make sense anyway so I think this is a reasonable fix.